refactor

+++ football_functions/DTOs/TeamDTO.cs
@@ -4,6 +4,6 @@ using System.Collections.Generic;
 namespace football_functions.DTOs;
 
 
-public readonly record struct TeamDTO(double Score, List<PlayerDTO> Players);
+public readonly record struct TeamDTO(decimal Score, List<PlayerDTO> Players);
